Move versions to shared
